Compute benchmark averages for all hosts
The compute_averages.py script is used to calculate the average latency, bandwidth and IOPS for all of the data that has been parsed from the benchmarks. In order to run properly, the 'parse_data' script must be run and saved to an output folder. This output folder must be specified as an argument while executing compute_averages.py.